Data_Preparation.ipynb now is working. NB. Paperspace truncate large files so KG were loaded with wrong shape
